Added arguments to mapper.
This is a backward-incompatible change that changes the signature
of all mappers from (row) to (row, ...args), and Map(name, f) to
Map(name, f, ...args) where args are passed by the driver to
the executors.

This allows for more complex map operations where the maps depend
on intermediary operations / arguments.